Add files via upload
File includes the pc/ec classes, as well as the main program and its relevant functions. The program initially prompts for the loading of characters, which I left blank because I believe that is someone's designated job to figure that out. Regardless of whether characters are loaded or not, the program also prompts for the creation of characters and if yes, how many. I already created functions and implemented that part of the code. Following this character initialization part is the start of the game where the user will enter commands from the provided list. I included notes on what to do when adding commands for when y'all do, so check those out. The basic structure for the if blocks are simple, so if you just take a peek at those, it should be easy to figure out. Good luck!